Reorganize the directory structure to match typical projects
