+++ index.html
@@ -11,7 +11,7 @@
   <body>
     <div id="app"></div>
     <script type="module">
-      import ChooseArea from "./ChooseArea.svelte";
+      import ChooseArea from "./src/ChooseArea.svelte";
 
       new ChooseArea({
         target: document.getElementById("app"),
