+++ package.json
@@ -7,7 +7,7 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "fmt": "npx prettier --write *.html *.js *.ts *.svelte components/* tests/*",
+    "fmt": "npx prettier --write *.html tests/* src/*",
     "lint": "npx eslint . --ignore-pattern=vite.config.js",
     "test": "npx playwright test",
     "check": "svelte-check --tsconfig ./tsconfig.json"
