+++ src/ChooseArea.svelte
@@ -2,10 +2,10 @@
   import maplibregl from "maplibre-gl";
   import "maplibre-gl/dist/maplibre-gl.css";
   import geojsonExtent from "@mapbox/geojson-extent";
-  import About from "./components/About.svelte";
+  import About from "./lib/About.svelte";
   import "carbon-components-svelte/css/white.css";
 
-  import authoritiesUrl from "./authorities.geojson?url";
+  import authoritiesUrl from "../assets/authorities.geojson?url";
 
   let showAbout = false;
