+++ src/lib/Legend.svelte
