+++ src/lib/ZoomOutMap.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import geojsonExtent from "@mapbox/geojson-extent";
   import { map } from "../stores.js";
-  import icon from "../assets/zoom_out_map.svg";
+  import icon from "../../assets/zoom_out_map.svg";
 
   export let boundaryGeojson;
