+++ src/lib/draw/AttributeMode.svelte
