+++ src/lib/draw/GeometryMode.svelte
