+++ src/lib/draw/HoverLayer.svelte
