+++ src/lib/draw/InterventionLayer.svelte
