+++ src/lib/draw/PointMode.svelte
