+++ src/lib/draw/PolygonMode.svelte
