+++ src/lib/draw/SplitRouteMode.svelte
