+++ src/lib/draw/point_tool.js
