+++ src/lib/draw/polygon_tool.js
